Исправление расширения карты
