+++ Zilon.Core/Zilon.Core/MapGenerators/MapFactoryHelper.cs
@@ -1,7 +1,7 @@
 ﻿using System;
 using System.Collections.Generic;
 using System.Linq;
-
+using Zilon.Core.Common;
 using Zilon.Core.Schemes;
 using Zilon.Core.World;
 
@@ -32,6 +32,34 @@ namespace Zilon.Core.MapGenerators
             var next = sectorNode.Biome.GetNext(sectorNode);
 
             return next.Select(node => new RoomTransition(node as ISectorNode));
+        }
+
+        public static Matrix<bool> ResizeMatrixTo7(Matrix<bool> matrix)
+        {
+            if (matrix is null)
+            {
+                throw new ArgumentNullException(nameof(matrix));
+            }
+
+            var resizedMatrix = matrix.CreateMatrixWithMargins(1, 1);
+            for (var x = 0; x < matrix.Width; x++)
+            {
+                for (var y = 0; y < matrix.Height; y++)
+                {
+                    if (matrix[x, y])
+                    {
+                         var neighbors = HexHelper.GetNeighbors(x + 1, y + 1);
+                        foreach (var neightbor in neighbors)
+                        {
+                            var resizedX = neightbor.X;
+                            var resizedY = neightbor.Y;
+                            resizedMatrix[resizedX, resizedY] = true;
+                        }
+                    }
+                }
+            }
+
+            return resizedMatrix;
         }
     }
 }
